Fix compiler warnings
* static_cast for signed/unsigned integer comparisons
* remove unused variables
* fix order of member initialization
* use override instead of virtual
* printf %d -> %ld
* clang pragmas

+++ src/lib/OpenEXRCore/internal_huf.c
@@ -1722,7 +1722,7 @@ fasthuf_decode (
         // 8 bits of data in the buffer
         //
 
-        if (symbol == rleSym)
+        if (symbol == (int) rleSym)
         {
             uint32_t rleCount;
